Following and leading words
The AutoCorrect dictionary now also keeps track of words that tend to
follow a given word or words that a word tends to lead. However, in this
commit this information is not yet used in any way.

Also made some changes to the overall structure of autocorrect.py Moved
some methods out of the main class and renamed some interinally used
methods.