Find similar words ratings
The method that finds similar words given a word can now use the following
and leading words introduced in the previous commit to determine a rating
for words similar to the given word. This gives a better idea of how
likely it is that a given similar word was the intended word.