Improved correcting of texts
Improved the .correct_text method, it should now be able to deal with
capital letters correctly as well as handle punctiation.

Also .find_similar_words now increases the rating of a suggestion if that
suggestion is the given word (a test for this is added as well).
Something else regarding the ratings, these are now defined in variables
at the top of AutoCorrect.py

And there have also been a few small changes reducing duplicate code and
improving code readability.

And last but not least, the default alphabet no includes apostrophe.