Improved BubbleSearch
Significantly improved the efficiency of BubbleSearch. Previously,
bubblesearch would swap letters around in the input string and search the
dictionary tree from start to finish each time. This very inefficient
algorithm has been replaced with a depth-first search in the tree itself.
By swapping letters in the string of the word, and checking if there
exists a node for the next letter in the string as a child of the current
node, this algorithm reduces the search space significantly (because it
automatically backtracks when it hits a dead end).