Improved MissingSearch
Significantly improved the MissingSearch algorithm by adding using the
dictionary tree instead of strings while exploring the possibility space.
This way the algorithm can detect early outs earlier (i.e., it finds out
that a node has no children for the next letter in the word) and it has a
reduced search space by only inserting letters that are a child of the
current node, instead of all letters in the alphabet.

Also added an internal helper method the get the child node of a node for
a particular letter, which is now used in both the BubbleSearch and
MissingSearch algorithm.