Minor improvements
Made some very small improvements to the overall module. Decreasing all
(but one) lines to have a maximum length of 80. Removed the `if node is
None` checks at the start of BubbleSearch and MissingSearch. Also made a
bigger improvement to BubbleSearch with an additional early out (with `if
node is None` in the for-loop) and as a result less if-statements after
the for-loop.

Also updated the last revision date in the ./AutoCorrect/autocorrect.py
and updated the years-range in the LICENSE file.