Add API to the README
Updated ./README.md to include the API described in ./docs/Dictionary.md
and removed the ./docs folder. Also merge the `.learn_file()` and
`.learn_text()` methods into one general `.learn_text()` method which
function similar to `.correct_text()` (where the parameters can be used to
correct a string of text or the text from a file).