Improved SpaceSearch
Improved the SpaceSearch algorithm by adding early outs. It now uses the
dictionary tree in real time so it can stop as soon as it turns out the
substring doesn't exist/isn't known. Similarly it will continue to the
next loop-iteration when it turns out the last node of the first substring
does not indicate that it is a word (`node['is_word']`).